(docs) ref: organisation with creation of a file explorer component
