+++ src/app/(user)/(cdp)/[study]/docs/page.tsx
@@ -1,5 +1,5 @@
 import { StudyParams } from '@/routes';
-import { DocumentList } from './doc_list';
+import { DocumentList } from './doc-list';
 import { getMissionFiles } from '@/drive/files';
 
 export default async function Page({ params }: StudyParams) {
